new feature: top up to wallet, rent pitch with wallet, get transactions, user response edited

+++ BE/httc-sport/src/main/java/com/fpoly/httc_sport/entity/Wallet.java
@@ -17,7 +17,7 @@ public class Wallet {
 	@Id
 	@GeneratedValue(strategy = GenerationType.UUID)
 	String id;
-	long money;
+	int money;
 	
 	@OneToOne
 	@JoinColumn(name = "user_id")
